daemon/requirements.txt
PyYAML>=6.0
PyMySQL>=1.1
